+++ packages/services/api/src/modules/organization/resolvers.ts
@@ -8,7 +8,6 @@ import {
 } from '../auth/providers/organization-access';
 import { isProjectScope, ProjectAccessScope } from '../auth/providers/project-access';
 import { isTargetScope, TargetAccessScope } from '../auth/providers/target-access';
-import { InMemoryRateLimiter } from '../rate-limit/providers/in-memory-rate-limiter';
 import { IdTranslator } from '../shared/providers/id-translator';
 import { Logger } from '../shared/providers/logger';
 import type { OrganizationModule } from './__generated__/types';
@@ -38,36 +37,6 @@ const createOrUpdateMemberRoleInputSchema = z.object({
 
 export const resolvers: OrganizationModule.Resolvers = {
   Mutation: {
-    async inviteToOrganizationByEmail(_, { input }, { injector }) {
-      await injector.get(InMemoryRateLimiter).check(
-        'inviteToOrganizationByEmail',
-        5_000, // 5 seconds
-        6, // 6 invites
-        `Exceeded rate limit for inviting to organization by email.`,
-      );
-
-      const InputModel = z.object({
-        email: z.string().email().max(128, 'Email must be at most 128 characters long'),
-      });
-      const result = InputModel.safeParse(input);
-
-      if (!result.success) {
-        return {
-          error: {
-            message: 'Please check your input.',
-            inputErrors: {
-              email: result.error.formErrors.fieldErrors.email?.[0],
-            },
-          },
-        };
-      }
-      const organization = await injector.get(IdTranslator).translateOrganizationId(input);
-      return await injector.get(OrganizationManager).inviteByEmail({
-        organization,
-        email: input.email,
-        role: input.role,
-      });
-    },
     async requestOrganizationTransfer(_, { input }, { injector }) {
       const organization = await injector.get(IdTranslator).translateOrganizationId(input);
       return injector.get(OrganizationManager).requestOwnershipTransfer({
